WIP: split CourseOfferDialog into two components: CourseDialog and CourseDialogForm to make the codebase more predictable. WIP: Also make CourseDialog more generic so you can use it to offer and edit courses.
